+++ setup.cfg
@@ -29,6 +29,4 @@ install_requires =
     uvicorn
     starlette
     slugid
-    fuse
-    simple-httpfs
 python_requires = >=3.7
